refactor(goal-detail): split the mobile goal-detail dialogs out of GoalDetailSheet (#467)
The last #467 split. Move GoalDetailSheet's five dialog/sheet components —
GoalActionsSheet, DeleteGoalConfirmSheet, EditGoalSheet, InvestmentActionSheet, and
UnassignConfirmSheet — into a new `goalDetailDialogs.tsx`. They're module-level and
prop-driven, so this is a relocation: GoalDetailSheet keeps the data + orchestration
(tabs, the sell/withdraw + maturity-resolve sheets, the transaction history) and
passes handlers down. The edit-goal write still goes through goalActions.updateGoal.
No behavior or markup change; the now-unused imports (several lucide icons, the
goalDetailShared UI strips, needs*MaturityAction, updateGoal) drop out of the view.

Parity: the GoalDetailSheet component tests + the goals E2E (mobile goal-detail
delete-requires-confirmation and edit-from-panel flows, which drive the actions /
delete / edit dialogs).

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
git diff --check clean; E2E goals (6) + fund-unassign (3) passed. GoalDetailSheet
1409 → 744 lines.
